Add annotation to prevent deletion of propagated SA and RB
When this annotation is present, service accounts and role bindings that
are propagated from the root namespace will not be deleted when the root copy is deleted.

The annotation is cloudfoundry.org/propagate-deletion="false"

+++ controllers/controllers/workloads/cfspace_controller.go
@@ -309,6 +309,10 @@ func (r *CFSpaceReconciler) reconcileServiceAccounts(ctx context.Context, space 
 
 	for index := range propagatedServiceAccounts.Items {
 		propagatedServiceAccount := propagatedServiceAccounts.Items[index]
+		if propagatedServiceAccount.Annotations[korifiv1alpha1.PropagateDeletionAnnotation] == "false" {
+			continue
+		}
+
 		if _, found := serviceAccountMap[propagatedServiceAccount.Name]; !found {
 			err = r.client.Delete(ctx, &propagatedServiceAccount)
 			if err != nil {
